+++ source/common/upstream/maglev_lb.h
@@ -34,7 +34,8 @@ class MaglevTable : public ThreadAwareLoadBalancerBase::HashingLoadBalancer,
                     Logger::Loggable<Logger::Id::upstream> {
 public:
   MaglevTable(const NormalizedHostWeightVector& normalized_host_weights,
-              double max_normalized_weight, uint64_t table_size, MaglevLoadBalancerStats& stats);
+              double max_normalized_weight, uint64_t table_size, bool use_logical_name,
+              MaglevLoadBalancerStats& stats);
 
   // ThreadAwareLoadBalancerBase::HashingLoadBalancer
   HostConstSharedPtr chooseHost(uint64_t hash) const override;
@@ -70,6 +71,7 @@ class MaglevLoadBalancer : public ThreadAwareLoadBalancerBase {
 public:
   MaglevLoadBalancer(const PrioritySet& priority_set, ClusterStats& stats, Stats::Scope& scope,
                      Runtime::Loader& runtime, Runtime::RandomGenerator& random,
+                     const absl::optional<envoy::api::v2::Cluster::MaglevLbConfig>& config,
                      const envoy::api::v2::Cluster::CommonLbConfig& common_config,
                      uint64_t table_size = MaglevTable::DefaultTableSize);
 
@@ -81,7 +83,7 @@ private:
   createLoadBalancer(const NormalizedHostWeightVector& normalized_host_weights,
                      double /* min_normalized_weight */, double max_normalized_weight) override {
     return std::make_shared<MaglevTable>(normalized_host_weights, max_normalized_weight,
-                                         table_size_, stats_);
+                                         table_size_, use_logical_name_, stats_);
   }
 
   static MaglevLoadBalancerStats generateStats(Stats::Scope& scope);
@@ -89,6 +91,7 @@ private:
   Stats::ScopePtr scope_;
   MaglevLoadBalancerStats stats_;
   const uint64_t table_size_;
+  const bool use_logical_name_;
 };
 
 } // namespace Upstream