+++ test/common/upstream/load_balancer_benchmark.cc
@@ -135,10 +135,12 @@ class MaglevTester : public BaseTester {
 public:
   MaglevTester(uint64_t num_hosts, uint32_t weighted_subset_percent = 0, uint32_t weight = 0)
       : BaseTester(num_hosts, weighted_subset_percent, weight) {
+    config_ = envoy::api::v2::Cluster::MaglevLbConfig();
     maglev_lb_ = std::make_unique<MaglevLoadBalancer>(priority_set_, stats_, stats_store_, runtime_,
-                                                      random_, common_config_);
+                                                      random_, config_, common_config_);
   }
 
+  absl::optional<envoy::api::v2::Cluster::MaglevLbConfig> config_;
   std::unique_ptr<MaglevLoadBalancer> maglev_lb_;
 };